Role Based Authorization

+++ convex/files.ts
@@ -32,7 +32,8 @@ async function hasAccessToOrg(ctx: QueryCtx | MutationCtx, orgId: string) {
   }
 
   const hasAccess =
-    user.orgIds.includes(orgId) || user.tokenIdentifier.includes(orgId)
+    user.orgIds.some((item) => item.orgId === orgId) ||
+    user.tokenIdentifier.includes(orgId)
 
   if (!hasAccess) {
     return null
@@ -111,6 +112,12 @@ export const deleteFile = mutation({
     const access = await hasAccessToFile(ctx, args.fileId)
 
     if (!access) throw new ConvexError('You do not have access to file')
+
+    const isAdmin =
+      access.user.orgIds.find((org) => org.orgId === access.file.orgId)
+        ?.role === 'admin'
+
+    if (!isAdmin) throw new ConvexError('You have no admin access')
 
     await ctx.db.delete(args.fileId)
   },
